requirements.txt
celery[redis]
Django == 1.11.2

#configure celery backend be django-db
django-celery-results

#database-backed periodic tasks with admin interface
django-celery-beat

ipython


Fabric3
fabtools
